+++ python/mxnet/ndarray/__init__.py
@@ -22,6 +22,4 @@ from .op import CachedOp
 # pylint: disable=wildcard-import, redefined-builtin
 from .ndarray import *
 from .utils import load, save, zeros, empty, array
-from .sparse_ndarray import _ndarray_cls
-#, csr_matrix, row_sparse_array
-#from .sparse_ndarray import BaseSparseNDArray, RowSparseNDArray, CSRNDArray
+from .sparse import _ndarray_cls
