+++ python/mxnet/symbol/symbol.py
@@ -2209,4 +2209,3 @@ def arange(start, stop=None, step=1.0, repeat=1, name=None, dtype=None):
                              name=name, dtype=dtype)
 
 _set_symbol_class(Symbol)
-
