+++ src/operator/tensor/sparse_retain.cc
@@ -41,14 +41,14 @@ Example::
   shape = (4, 2)
   rsp_in = row_sparse(data, indices)
   to_retain = [0, 3]
-  rsp_out = _sparse_retain(rsp_in, to_retain)
+  rsp_out = retain(rsp_in, to_retain)
   rsp_out.values = [[1, 2], [5, 6]]
   rsp_out.indices = [0, 3]
 
-The storage type of ``_sparse_retain`` output depends on storage types of inputs
+The storage type of ``retain`` output depends on storage types of inputs
 
-- _sparse_retain(row_sparse, default) = row_sparse
-- otherwise, ``_sparse_retain`` is not supported
+- retain(row_sparse, default) = row_sparse
+- otherwise, ``retain`` is not supported
 
 )code" ADD_FILELINE)
 .set_num_inputs(2)
